Move SQS health indicator to actuator module.
