+++ spring-cloud-starter-aws/pom.xml
@@ -43,5 +43,9 @@
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-aws-autoconfigure</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework.cloud</groupId>
+			<artifactId>spring-cloud-aws-actuator</artifactId>
+		</dependency>
 	</dependencies>
 </project>
